+++ registryGenerator.h
@@ -1,6 +1,5 @@
 #include <stdbool.h>
 #include <stdio.h>
-#include "libs/chainedList/chainedList.h"
 #include "registry.h"
 
 #ifndef _REGISTRY_GENERATOR_
@@ -11,7 +10,7 @@ Registry *initRandomRegistry();
 
 void writeRegistryAtFile(FILE *file, Registry *registry);
 
-void writeListInFile(unsigned int howMany);
+void writeListInFile(double howMany);
 
 int getSizeOfRegistry();
